khanh/sua file_route

+++ app/routes/file_route.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, HTTPException, Depends, status
 from app.schemas.api_response import APIResponse
-from app.schemas.file_schema import FileCreate, FileResponse
+from app.schemas.file_schema import FileCreate, FileResponse, DeleteListDocumentsRequest, FileDeleteResponse
 from sqlalchemy.orm import Session
 from app.core.database import get_db
 from app.models.document_model import Document
@@ -9,51 +9,47 @@ from app.schemas.folder_schema import FolderResponse
 from ..schemas.token_schema import TokenData
 from ..utils.oauth2 import get_current_user
 
+
 router = APIRouter(
-    prefix="/file",
+    prefix="/identity/api/v1/file",
     tags=['File']
 )
 
-# API: Lấy file theo slug
+# API: Get file by slug
 @router.get("/{slug}", response_model=APIResponse)
 def get_file_by_slug(slug: str, current_user: TokenData = Depends(get_current_user), db: Session = Depends(get_db)):
     folder = db.query(Folder).filter(Folder.slug == slug).first()
 
     if not folder:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail={"code": 1002, "message": "Folder không tồn tại"})
+                            detail={"code": 1002, "message": "Folder does not exist"})
 
-    documents = db.query(Document).filter(Document.folder_id == folder.id).all()
-
-    if not documents:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail={"code": 1003, "message": "Không có tài liệu", "error_message": "Không tìm thấy tài liệu trong folder"})
+    documents = db.query(Document).filter(Document.folder_id == folder.id).order_by(Document.id.desc()).all()
 
     items = []
     for document in documents:
         items.append({
             "id": str(document.id),
             "name": document.name,
-            "firebaseId": document.firebase_id,
-            "createAt": document.create_at,
+            "fire_base_id": document.firebase_id,
+            "create_at": document.create_at,
             "folder": {
                 "id": str(folder.id),
                 "name": folder.name,
                 "view": folder.view,
                 "star": folder.star,
-                "createAt": folder.create_at,
+                "create_at": folder.create_at,
                 "slug": folder.slug,
                 "author": {
                     "id": str(folder.author.id),
                     "username": folder.author.username,
                     "firstName": folder.author.first_name,
-                    "lastName": folder.author.last_name,
+                    "last_name": folder.author.last_name,
                     "dob": folder.author.dob,
                     "picture": folder.author.picture,
                     "location": folder.author.location,
                     "phone": folder.author.phone,
                     "email": folder.author.email,
-                    "noPassword": folder.author.no_password,
                     "activated": folder.author.activated,
                 }
             }
@@ -64,11 +60,11 @@ def get_file_by_slug(slug: str, current_user: TokenData = Depends(get_current_us
         result={
             "items": items,
             "total": len(items),
-            "canUpdate": True
+            "can_update": True
         }
     )
 
-# API: Thêm file vào folder
+# API: Add file to folder
 @router.post("/add/{slug}", response_model=APIResponse)
 def add_file(slug: str, file_data: FileCreate,
              current_user: TokenData = Depends(get_current_user),
@@ -77,9 +73,22 @@ def add_file(slug: str, file_data: FileCreate,
 
     if not folder:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail={"code": 1002, "message": "Folder không tồn tại"})
+                            detail={"code": 1002, "message": "Folder does not exist"})
 
-    new_file = Document(name=file_data.name, firebase_id=file_data.firebase_id, folder_id=folder.id)
+    if folder.author_id != current_user.user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail={"code": 1004, "message": "You are not authorized to add files to this folder"})
+
+    existing_file = db.query(Document).filter(Document.firebase_id == file_data.firebase_id).first()
+    if existing_file:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,
+                            detail={"code": 1003, "message": "firebase_id already exists in the documents"})
+
+    new_file = Document(
+        name=file_data.name,
+        firebase_id=file_data.firebase_id,
+        folder_id=folder.id
+    )
 
     db.add(new_file)
     db.commit()
@@ -106,14 +115,16 @@ def add_file(slug: str, file_data: FileCreate,
         )
     )
 
-# API: Lấy document theo id
+
+
+# API: Get document by id
 @router.get("/document/{id}", response_model=APIResponse)
 def get_document_by_id(id: int, current_user: TokenData = Depends(get_current_user), db: Session = Depends(get_db)):
     document = db.query(Document).filter(Document.id == id).first()
 
     if not document:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail={"code": 1002, "message": "Document không tồn tại"})
+                            detail={"code": 1002, "message": "Document does not exist"})
 
     return APIResponse(
         code=1000,
@@ -134,20 +145,51 @@ def get_document_by_id(id: int, current_user: TokenData = Depends(get_current_us
         )
     )
 
-# API: Xóa document theo id
+# API: Delete document by id
 @router.delete("/document/{id}")
 def delete_document_by_id(id: int, current_user: TokenData = Depends(get_current_user), db: Session = Depends(get_db)):
     document = db.query(Document).filter(Document.id == id).first()
+
     if not document:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail={"code": 1002, "message": "Document không tồn tại"})
+                            detail={"code": 1002, "message": "Document does not exist"})
+
+    folder = db.query(Folder).filter(Folder.id == document.folder_id).first()
+    if folder.author_id != current_user.user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail={"code": 1004, "message": "You are not authorized to delete this document"})
+
     db.delete(document)
     db.commit()
-    return {"message": "Document đã được xóa"}
 
-# API: Xóa tất cả các document
+    return {"message": "Document has been deleted"}
+
+
+# API: Delete documents
 @router.delete("/document")
-def delete_all_documents(current_user: TokenData = Depends(get_current_user), db: Session = Depends(get_db)):
-    db.query(Document).delete()
-    db.commit()
-    return {"message": "Tất cả document đã được xóa"}
+def delete_list_documents(delete_list_documents_request: DeleteListDocumentsRequest,
+                          current_user: TokenData = Depends(get_current_user),
+                          db: Session = Depends(get_db)):
+    document_ids = delete_list_documents_request.documents_id
+
+    if len(document_ids) > 0:
+        for doc_id in document_ids:
+            document = db.query(Document).filter(Document.id == doc_id).first()
+            if not document:
+                raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                                    detail={"code": 1002, "message": f"Document with id {doc_id} not found"})
+
+            folder = db.query(Folder).filter(Folder.id == document.folder_id).first()
+            if folder.author_id != current_user.user_id:
+                raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                                    detail={"code": 1004,
+                                            "message": "You can't delete documents not in your own folder"})
+
+            db.delete(document)
+
+        db.commit()
+
+    return APIResponse(
+        code=1000,
+        result=FileDeleteResponse(result="Delete successfully")
+    )
